+++ src/transport/virtual_device/builders.rs
@@ -35,6 +35,7 @@ pub(super) fn build_device_info(state: &VirtualDeviceState) -> Vec<u8> {
         OperationCode::GetObjectInfo.into(),
         OperationCode::GetObject.into(),
         OperationCode::GetPartialObject.into(),
+        OperationCode::GetPartialObject64.into(),
         OperationCode::GetThumb.into(),
         OperationCode::SendObjectInfo.into(),
         OperationCode::SendObject.into(),